web/apps/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.0
djangorestframework==3.14.0
kombu==5.3.1
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
